Break out emulator test fixtures
